endpoint: Move id allocation out of pkg/endpoint/id
All users of pkg/endpoint/id except for one are using the package to use
helpers around endpoint IDs. The ID allocation pulls in additional
dependencies, split them apart and move the id allocation into
pkg/endpointmanager where it is used.
