+++ pkg/endpointmanager/idallocator/allocator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package id
+package idallocator
 
 import (
 	"fmt"
